docs(SideBar): update api doc
