+++ packages/scripts/api.json
@@ -100798,7 +100798,7 @@
       ],
       "field_default_value": "",
       "field_enum": "",
-      "field_desc_zh": "`1.0.0-rc.1`。图标，传对象则透传至 Icon",
+      "field_desc_zh": "图标，传对象则透传至 Icon",
       "field_desc_en": null,
       "field_required": 0,
       "event_input": "",
